Moved checks and helpers to common.py
Fixes #87
